feat: JEI 配方背景使用机器纹理 + Canner 流体渲染 + 新增流体单元
- 所有机器 JEI RecipeCategory 改用机器 GUI 纹理作为背景
- 新增施工泡沫/蒸汽/过热蒸汽流体单元物品
- Canner 混合配方: 流体槽渲染使用 JEI 流体 API, 隐藏物品槽用于 R 查找
- 洗矿机: 水消耗渲染为流体槽
- 新增 JEI 裁取预览工具 (tools/JeiCropPreview.java)
- 修复部分渲染器冗余类型转换
- 移除不必要的 client 源集插件, 全部合并到 main 源集
- 移除本地 fabric-guidebook 复合构建, 使用 JitPack 依赖

+++ core/src/client/kotlin/ic2_120/client/renderers/TeleporterBlockEntityRenderer.kt
@@ -122,7 +122,7 @@ class TeleporterBlockEntityRenderer(
         drawRingBand(matrices, vc, fullLight, overlay, radiusOuter * 0.68f, radiusInner * 0.58f, ringY + 0.03f, 120, 250, 255, alpha - 20)
 
         if (progress > 0.7f) {
-            val beamHeight = (target.height + 0.6f).toFloat()
+            val beamHeight = target.height + 0.6f
             drawBeam(matrices, vc, fullLight, overlay, beamHeight, 70, 220, 255, (alpha + 20).coerceAtMost(230))
         }
